backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
google-generativeai
